requirements.txt

# Core web framework
Flask==2.3.3

# Forms & auth
Flask-WTF==1.2.2
Flask-Login==0.6.3
WTForms==3.0.1
email-validator==2.3.0

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
Flask-Migrate==4.0.4

# Templating & utilities
Jinja2==3.1.2
Werkzeug==2.2.3
itsdangerous==2.1.2
python-dotenv==1.0.0

# Data processing
pandas==2.1.2
openpyxl==3.1.2
xlrd==2.0.1

# Mail and ASGI adapter for Vercel
Flask-Mail==0.9.1
asgiref==3.8.0
psycopg2-binary==2.9.10

